% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_series.R
\name{animate_series}
\alias{animate_series}
\title{Animate a time series of rasters}
\usage{
animate_series(r, dates, breaks, param = NULL, param.name = "Parameter",
  file.name = tempfile(fileext = ".gif"), ...)
}
\arguments{
\item{r}{raster or stack of rasters, each representing a different acquisition time.}

\item{dates}{character vector of length \code{nlayers(r)}, representing corresponding times of \code{r}.}

\item{breaks}{numeric vector, value range to be displayed, e.g. \code{seq(1, 180, by = 1)}}

\item{param}{data.frame or \code{NULL}, optional parameter data.frame derived using \link{bands_param} to show the development of a paramater over time.}

\item{param.name}{character, name of the defined paramater. Default is "Parameter".}

\item{file.name}{character, path to the output file Default is a temporary file.}

\item{...}{arguments passed to \link{saveGIF}, e.g. interval=0.2, ani.width = 500, ani.height = 700 etc.}
}
\value{
List of plots that are used as frames for the animation. An animation file will be written to file.name.
}
\description{
\code{animate_series} animates a series of rasters in consecutive order.
}
