% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_param.R
\name{band_param}
\alias{band_param}
\title{Calculate a band paramater}
\usage{
band_param(r, fun)
}
\arguments{
\item{r}{raster or stack of rasters, each representing a different acquisition time.}

\item{fun}{function to be applied.}
}
\value{
Two-column data.frame with "stat" containing the band parameter(s) and "time" indicating consecutive order of the parameter in time
}
\description{
\code{scene_param} applies a function to all cell values per band to retrieve a band parameter.
}
