% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Dummy function}
\usage{
dummy(x, y, additive)
}
\arguments{
\item{x}{numeric, a number of your choice.}

\item{y}{numeric, a number to be multiplied-}

\item{additive}{logical, if \code{TRUE}, x and y are added, else they are substracted..}
}
\value{
Describe, what is returned by your function, for example: a numeric value \code{z}.
}
\description{
A short description of your function, for example: \code{dummy} is a dummy function,
serving as a template for you  to create new functions. Just copy this file into the
R directory of this package, rename it to the name of your function and start changing 
it, including this text.
}
