% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoTap_scenario_GT2013.R
\name{likelihood_2Tap_EqVar}
\alias{likelihood_2Tap_EqVar}
\title{Compute likelihood for the two-tap scenario with equal spatial variance}
\usage{
likelihood_2Tap_EqVar(x1m, x2m, x1, x2, sigma_s)
}
\arguments{
\item{x1m}{measured/sensed or hypothetical position of tap 1 for which the likelihood (p given the other parameters) is computed}

\item{x2m}{measured/sensed or hypothetical position of tap 2}

\item{x1}{(real) position of tap 1}

\item{x2}{(real) position of tap 1}

\item{sigma_s}{spatial uncertainty (given as a standard deviation) for the taps (same sd assumed for both positions)}
}
\value{
likelihood values (see article for details or the vignette on the two-tap scenario). If x1m and x2m are vectors, a vector of likelihood values is returned.
}
\description{
The function computes likelihood values for the given trajectory.
The code refers to the two-tap model with equal variance (spatial uncertainty) for both taps as described in Goldreich & Tong (2013), given in formula A11.
}
\examples{
require(rabBITS)

#XXX Example 1: compute a single point estimate XXX
likelihood_2Tap_EqVar(x1m=2, x2m=4, x1=1, x2=2, sigma_s=1)

#XXX Example 2: plot a prior distribution for cobinations of tap 1 and 2 XXX
library(ggplot2)
x1m_range <- c(0, 10) #range for taps
x2m_range <- c(0, 10)

x1m_res <- 100 #resolution for graphs
x2m_res <- 100

likelihoodMat <- expand.grid(x1m=seq(x1m_range[1], x1m_range[2], length.out = x1m_res), x2m=seq(x2m_range[1], x2m_range[2], length.out = x2m_res))
likelihoodMat$l <- likelihood_2Tap_EqVar(x1m = likelihoodMat$x1m, x2m = likelihoodMat$x2m, x1=2, x2=4, sigma_s = 1)
ggplot(likelihoodMat, aes(x=x1m, y=x2m, fill=l)) +
  geom_raster() +
  coord_fixed() +
  ggtitle("likelihood")

}
\references{
Goldreich & Tong, 2013, Frontiers in Psychology
}
