% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoTap_equalVar_GT2013.R
\name{posterior_params_2Tap_EqVar}
\alias{posterior_params_2Tap_EqVar}
\title{Compute parameters of the posterior distribution for the two-tap scenario with equal spatial variance}
\usage{
posterior_params_2Tap_EqVar(x1m, x2m, time_t, sigma_s, sigma_v)
}
\arguments{
\item{x1m}{measured/sensed or hypothetical position of tap 1 for which the likelihood (p given the other parameters) is computed}

\item{x2m}{measured/sensed or hypothetical position of tap 2}

\item{time_t}{speed prior (in units of space per time; given as a standard deviation)}

\item{sigma_s}{spatial uncertainty (given as a standard deviation) for the taps (same sd assumed for both positions)}

\item{sigma_v}{speed prior (in units of space per time; given as a standard deviation)}
}
\value{
a list containing the posterior modes, the variance, and the correlation.
}
\description{
The function computes parameters of the 2-dimensional posterior distrbution, namely, the posterior modes for both dimensions (corresponding to taps 1 and tap 2),
the variance, and correlation.
The code refers to the two-tap model with equal variance (spatial uncertainty) for both taps as described in Goldreich & Tong (2013), given in \strong{Formula A14}.
}
\examples{
#posterior parameters for the given parameters...
posterior_params_2Tap_EqVar(x1m=2, x2m=4, time_t=0.1, sigma_s=1, sigma_v=10)

}
\references{
Goldreich & Tong, 2013, Frontiers in Psychology
}
