% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoTap_scenario_GT2013.R
\name{prior_2Tap}
\alias{prior_2Tap}
\title{Compute probability density for the two-tap scenario}
\usage{
prior_2Tap(x1, x2, time_t, sigma_v)
}
\arguments{
\item{x1}{position of tap 1}

\item{x2}{position of tap 2}

\item{time_t}{time passing between tap 1 and 2}

\item{sigma_v}{speed prior (in units of space per time; given as a standard deviation)}
}
\value{
values that are proportional to the prior probability for the given trajectory. If x1 and x2 are vectors, a vector of prior probabilities.
Note that the returned values are not strictly speaking probabilities because they are not normalized (see article for details or the vignette on the two-tap scenario).
}
\description{
The function computes values that are proportinoal to the prior probability for a trajectory defined by two stimuli. The code refers
to the two-tap model as described in Goldreich & Tong (2013), given in formula A08.
}
\examples{
require(rabBITS)

#XXX Example 1: compute a single point estimate XXX
prior_2Tap(x1=2, x2=4, sigma_v=10, time_t=0.1)

#XXX Example 2: compute a whole vector of combinations of tap 1 and 2 XXX
prior_2Tap(x1=1, x2=c(1:10), sigma_v=10, time_t=0.1)

#XXX Example 3: plot a prior distribution for combinations of tap 1 and 2 XXX
library(ggplot2)
x1_range <- c(0, 10) #range for taps
x2_range <- c(0, 10)

x1_res <- 100 #resolution for graphs
x2_res <- 100

priorMat <- expand.grid(x1=seq(x1_range[1], x1_range[2], length.out = x1_res),
x2=seq(x2_range[1], x2_range[2], length.out = x2_res))
priorMat$p <- prior_2Tap(x1 = priorMat$x1, x2 = priorMat$x2, sigma_v = 10, time_t = 0.1)
ggplot(priorMat, aes(x=x1, y=x2, fill=p)) +
  geom_raster() +
  coord_fixed() +
  ggtitle("prior")

}
\references{
Goldreich & Tong, 2013, Frontiers in Psychology
}
