% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit2.R
\name{strsplit1}
\alias{strsplit1}
\title{split something}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{the thing to be split}

\item{split}{the things where it is split}
}
\value{
returns the splitted things
}
\description{
does somehting nice
}
\examples{
strsplit1("asda sdsa", split = " ")
}
